rename EventsFrame to HomeEventsFrame and update refs
