+++ frontend/components/HomeEventsFrame.tsx
